requirements.txt
tensorflow
tensorflow-hub
opencv-python
streamlit
numpy
pillow
watchdog